requirements.txt
ruamel.yaml==0.17.16
web3==5.24.0
requests~=2.28.2

pydantic~=2.3.0